bug hunt

+++ src/jit/run.c
@@ -3,14 +3,14 @@
 #include "compiler.h"
 #include "logging.h"
 
-typedef int (*func_ptr)();
+typedef int64_t (*func_ptr)();
 
-int neart_jit_exec(vmctx_t * ctx) {
+int64_t neart_jit_exec(vmctx_t * ctx) {
 
     rcode_t * main = ctx->code + ctx->main_offset;
 
     func_ptr f = (func_ptr)neart_jit_compile(ctx, main);
 
-    int result = f();
+    int64_t result = f();
     return result;
 }
